fixed image browsing bug
Formerly when an image was selected and one navigated to the next image in the gallery, only the image metadata would update but the image data URL would remain that of the initially selected image. To address this, the currently selected image data URL is stored in App.svelte, and the ImageData interface has been renamed to ImageMeta, and a new interface named ImageData is a superset of the old ImageData, adding the data url property.